Revert "Added the CR repo for a new release against Samba 4.10"
The repo existed but it does not provide source packages, so can't build
against it. At the same time, the previously built release for
Samba 4.8 CAN be used for the time being and it effectively makes
Windows clients FAIL (cf. CERNBOX-1072).
To be checked if the recent dummy code is needed at all, once this
package can be successfully compiled against Samba 4.10.

This reverts commit 42b7749df4f5ad3c5b727ac5bf6a1505ef16d123.